pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = 'qupyt'
version = '0.1.0'
description = 'Provide a framework for quantum sensing experiments'
license = {text = 'No license of for commercial or non-commercial use is granted'}
dependencies = [
    'numpy', 
    'scipy',
    'harvesters',
    'matplotlib',
    'nidaqmx',
    'numpy',
    'pulsestreamer',
    'pypylon',
    'pyserial',
    'PyVISA',
    'pyvisa-py',
    'PyYAML',
    'termcolor',
    'tqdm',
    'watchdog',
]
#author = 'Karl Briegel'
#author_email = 'karl.briegel@tum.de'

[tool.setuptools.packages]
find = {}

[project.scripts]
qupyt = "qupyt:entry"

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true

[tool.pylsp]
configurationSources = ["pyproject.toml"]

[tool.pylsp.plugins.pylint]
enabled = true
args = ["--disable=line-too-long"]
